- Changed max iter to 32 bit from 16 - Fixed issue with calc progress getting stuck

+++ source/mandelbrotcalc.h
@@ -38,7 +38,7 @@ public:
         double              y1;         // y coordinate bounding point 1
         double              x2;         // x coordinate bounding point 2
         double              y2;         // y coordinate bounding point 2
-        uint16_t            iter_mx;    // Max iterations
+        uint32_t            iter_mx;    // Max iterations
         uint16_t            th_cnt;     // Thread count
     };
 
@@ -53,11 +53,11 @@ public:
         double                  y1;         // y coordinate bounding point 1 (adjusted)
         double                  x2;         // x coordinate bounding point 2 (adjusted)
         double                  y2;         // y coordinate bounding point 2 (adjusted)
-        uint16_t                iter_mx;    // Max iterations
+        uint32_t                iter_mx;    // Max iterations
         uint16_t                th_cnt;     // Thread count used
         uint16_t                img_width;  // Image width
         uint16_t                img_height; // Imahe height
-        std::vector<uint16_t>   img_data;   // Image data (internal buffer)
+        std::vector<uint32_t>   img_data;   // Image data (internal buffer)
         uint64_t                time_ms;    // Calc time in milliseconds
     };
 
@@ -92,9 +92,8 @@ private:
     std::atomic<int32_t>        m_y_cur;            // Current image line to process (negative means no work)
     std::atomic<int32_t>        m_y_done;           // Completed image lines
     int32_t                     m_y_upd;
-    //uint32_t                    m_data_cur_size;    // Current image size
-    uint16_t *                  m_data;             // Image buffer
-    uint16_t                    m_mxi;              // Max iterations
+    uint32_t *                  m_data;             // Image buffer
+    uint32_t                    m_mxi;              // Max iterations
     uint16_t                    m_w;                // Image width
     uint16_t                    m_h;                // Image height
     double                      m_x1;               // Initial X value
@@ -104,7 +103,7 @@ private:
     bool                        m_exit;
 
     void controlThread();
-    void workerThread( uint8_t id );
+    void workerThread( uint16_t id );
 };
 
 #endif // MANDELBROTCALC_H
